+++ node/cli/src/benchmarking.rs
@@ -155,7 +155,7 @@ pub fn create_benchmark_extrinsic(
 	runtime::UncheckedExtrinsic::new_signed(
 		call.clone(),
 		sp_runtime::AccountId32::from(sender.public()).into(),
-		Signature::Sr25519(signature.clone()),
+		Signature::Sr25519(signature),
 		extra.clone(),
 	)
 }
